move more responsibilities to AcquisitionService
